Move media utils into project. Add comments. Fix image/video node.
